feat: implementação das camadas de dados e de serviço, incluindo:
- Criação de interface "interfaces_dados.hpp".
- Criação de diretórios "data" (MDPE e MDPR) e "service" (MSA, MSC e MSP).
- Implementação dos arquivos .hpp e .cpp relacionados a camada de dados.
- Implementação dos arquivos .hpp e .cpp relacionados a camada de serviço.
- Criação automática do banco de dados na raiz do projeto ao rodar pela primeira vez.
- Implementação de main que faz teste de criação de banco e inserção de dados (APENAS RETIRAR O BLOCO TRY/CATCH NO FUTURO E MANTER TODO O RESTANTE PARA GARANTIR A INJEÇÃO DE DEPENDÊNCIAS!!!).
- Entrega final dos artefatos relacionados a modelagem.
- Recriação de documentação doxygen que agora contém novas classe (interfaces e módulos).

+++ projetoTP1/dominios/dominios.cpp
@@ -85,7 +85,7 @@ void Data::setData(string dataFornecida) {
 
 // EMAIL
 void Email::validar(string emailFornecido) {
-    int posArroba = emailFornecido.find('@');
+    size_t posArroba = emailFornecido.find('@');
     if (posArroba == string::npos || emailFornecido.find('@', posArroba + 1) != string::npos) {
         throw invalid_argument("Email invalido: deve conter exatamente um @.");
     }
